+++ src/spirit/src/bfloat16.cpp
@@ -28,7 +28,7 @@ float bfloat16_to_float32(const uint16_t bfloat16)
         constexpr uint32_t exponent_bias = 127 + 7;  // 127: 指数部の0x7Fのオフセット, 7: 仮数部のビット数
         const uint32_t actual_mantissa = mantissa + 128;  // ケチ表現で省略された最上位ビットを復元
 
-        return std::ldexp(actual_mantissa, exponent - exponent_bias) * sign;
+        return (float)std::ldexp(actual_mantissa, exponent - exponent_bias) * sign;
     }
 }
 
@@ -54,7 +54,7 @@ uint16_t float32_to_bfloat16(const float value)
         mantissa                     = static_cast<uint32_t>(std::ldexp(mantissa_float, 8));
     }
 
-    return (sign << 15) | (exponent << 7) | mantissa;
+    return (uint16_t)((sign << 15) | (exponent << 7) | mantissa);
 }
 
 }  // namespace spirit
